Supress warnings from urllib3
urllib3 is used internally in the google-cloud-storage library.
urllib3 logs warnings when more than a configured number of
threads have a connection setup to the same source:
"Connection pool is full, discarding connection:
storage.googleapis.com". This just means that the connection
is thrown out of the connection pool after the process is finished,
which doesn't matter to us. We therefore supress the warning
by suppressing all warnings from this library and rely on
their errors and checks in our code to detect any problems.