Attempt to fix error in rna_align on larger files
sambamba-sort throws an error on a real RNA dataset.
'sambamba-sort: Error reading BGZF block starting from offset 0:
stream error: not enough data in stream'
It is possible that putting a 'sambamba view' command in front of
the 'sambamba sort' command gives it the input that it expects,
and it is worth a shot.